flask: the initial flask application layout for serving RRD files
This is a basic implementation of an RRD Javascript viewer stuck
in the framework of a Flask application. The combination of a
Flask + GraphQL + React + JS should allow a standalone python
environment to be used to serve the data without any dependency
on the CGI examples from the collectd front ends.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>