setup: update the requirements to rrdtool since that is current
Looked into the rrdtool 1.7.0 release and the setup.py in that
is called rrdtool and the pypi release site is showing that the
releases are coming in 2018 and are being updated.

So, change the requirements and setup to that rrdtool package
to track those changes. Make sure the version is at 0.1.15 to
avoid an invalid free the was occurring during test.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>